优化移植代码逻辑

+++ app/src/main/java/cn/edu/cqupt/dmb/player/tuner/exoplayer/audio/MediaCodecAudioDecoder.java
@@ -179,7 +179,7 @@ public class MediaCodecAudioDecoder extends AudioDecoder {
     private int getDecodeOnlyIndex(long presentationTimeUs) {
         final int size = mDecodeOnlyPresentationTimestamps.size();
         for (int i = 0; i < size; i++) {
-            if (mDecodeOnlyPresentationTimestamps.get(i).longValue() == presentationTimeUs) {
+            if (mDecodeOnlyPresentationTimestamps.get(i) == presentationTimeUs) {
                 return i;
             }
         }
